Rock-paper-scissor game
In this task, I created a "Rock-Paper-Scissors game using Python" , where the user can play against the computer. The program takes the user's input, generates a random move for the computer, and then displays the winner based on the game rules. It also tracks scores and allows multiple rounds of play.